@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel.localedata
     ~~~~~~~~~~~~~~~~
@@ -8,26 +7,30 @@
     :note: The `Locale` class, which uses this module under the hood, provides a
            more convenient interface for accessing the locale data.
 
-    :copyright: (c) 2013-2022 by the Babel Team.
+    :copyright: (c) 2013-2023 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
-import pickle
+from __future__ import annotations
+
 import os
+import pickle
 import re
 import sys
 import threading
 from collections import abc
+from collections.abc import Iterator, Mapping, MutableMapping
+from functools import lru_cache
 from itertools import chain
+from typing import Any
 
-
-_cache = {}
+_cache: dict[str, Any] = {}
 _cache_lock = threading.RLock()
 _dirname = os.path.join(os.path.dirname(__file__), 'locale-data')
 _windows_reserved_name_re = re.compile("^(con|prn|aux|nul|com[0-9]|lpt[0-9])$", re.I)
 
 
-def normalize_locale(name):
+def normalize_locale(name: str) -> str | None:
     """Normalize a locale ID by stripping spaces and apply proper casing.
 
     Returns the normalized locale ID string or `None` if the ID is not
@@ -41,7 +44,7 @@ def normalize_locale(name):
             return locale_id
 
 
-def resolve_locale_filename(name):
+def resolve_locale_filename(name: os.PathLike[str] | str) -> str:
     """
     Resolve a locale identifier to a `.dat` path on disk.
     """
@@ -51,13 +54,13 @@ def resolve_locale_filename(name):
 
     # Ensure we're not left with one of the Windows reserved names.
     if sys.platform == "win32" and _windows_reserved_name_re.match(os.path.splitext(name)[0]):
-        raise ValueError("Name %s is invalid on Windows" % name)
+        raise ValueError(f"Name {name} is invalid on Windows")
 
     # Build the path.
-    return os.path.join(_dirname, '%s.dat' % name)
+    return os.path.join(_dirname, f"{name}.dat")
 
 
-def exists(name):
+def exists(name: str) -> bool:
     """Check whether locale data is available for the given locale.
 
     Returns `True` if it exists, `False` otherwise.
@@ -72,31 +75,27 @@ def exists(name):
     return True if file_found else bool(normalize_locale(name))
 
 
-def locale_identifiers():
+@lru_cache(maxsize=None)
+def locale_identifiers() -> list[str]:
     """Return a list of all locale identifiers for which locale data is
     available.
 
-    This data is cached after the first invocation in `locale_identifiers.cache`.
-
-    Removing the `locale_identifiers.cache` attribute or setting it to `None`
-    will cause this function to re-read the list from disk.
+    This data is cached after the first invocation.
+    You can clear the cache by calling `locale_identifiers.cache_clear()`.
 
     .. versionadded:: 0.8.1
 
     :return: a list of locale identifiers (strings)
     """
-    data = getattr(locale_identifiers, 'cache', None)
-    if data is None:
-        locale_identifiers.cache = data = [
-            stem
-            for stem, extension in
-            (os.path.splitext(filename) for filename in os.listdir(_dirname))
-            if extension == '.dat' and stem != 'root'
-        ]
-    return data
+    return [
+        stem
+        for stem, extension in
+        (os.path.splitext(filename) for filename in os.listdir(_dirname))
+        if extension == '.dat' and stem != 'root'
+    ]
 
 
-def load(name, merge_inherited=True):
+def load(name: os.PathLike[str] | str, merge_inherited: bool = True) -> dict[str, Any]:
     """Load the locale data for the given locale.
 
     The locale data is a dictionary that contains much of the data defined by
@@ -119,7 +118,7 @@ def load(name, merge_inherited=True):
     :param merge_inherited: whether the inherited data should be merged into
                             the data of the requested locale
     :raise `IOError`: if no locale data file is found for the given locale
-                      identifer, or one of the locales it inherits from
+                      identifier, or one of the locales it inherits from
     """
     name = os.path.basename(name)
     _cache_lock.acquire()
@@ -134,10 +133,7 @@ def load(name, merge_inherited=True):
                 parent = get_global('parent_exceptions').get(name)
                 if not parent:
                     parts = name.split('_')
-                    if len(parts) == 1:
-                        parent = 'root'
-                    else:
-                        parent = '_'.join(parts[:-1])
+                    parent = "root" if len(parts) == 1 else "_".join(parts[:-1])
                 data = load(parent).copy()
             filename = resolve_locale_filename(name)
             with open(filename, 'rb') as fileobj:
@@ -151,7 +147,7 @@ def load(name, merge_inherited=True):
         _cache_lock.release()
 
 
-def merge(dict1, dict2):
+def merge(dict1: MutableMapping[Any, Any], dict2: Mapping[Any, Any]) -> None:
     """Merge the data from `dict2` into the `dict1` dictionary, making copies
     of nested dictionaries.
 
@@ -184,20 +180,20 @@ def merge(dict1, dict2):
             dict1[key] = val1
 
 
-class Alias(object):
+class Alias:
     """Representation of an alias in the locale data.
 
     An alias is a value that refers to some other part of the locale data,
     as specified by the `keys`.
     """
 
-    def __init__(self, keys):
+    def __init__(self, keys: tuple[str, ...]) -> None:
         self.keys = tuple(keys)
 
-    def __repr__(self):
-        return '<%s %r>' % (type(self).__name__, self.keys)
+    def __repr__(self) -> str:
+        return f"<{type(self).__name__} {self.keys!r}>"
 
-    def resolve(self, data):
+    def resolve(self, data: Mapping[str | int | None, Any]) -> Mapping[str | int | None, Any]:
         """Resolve the alias based on the given data.
 
         This is done recursively, so if one alias resolves to a second alias,
@@ -222,19 +218,19 @@ class LocaleDataDict(abc.MutableMapping):
     values.
     """
 
-    def __init__(self, data, base=None):
+    def __init__(self, data: MutableMapping[str | int | None, Any], base: Mapping[str | int | None, Any] | None = None):
         self._data = data
         if base is None:
             base = data
         self.base = base
 
-    def __len__(self):
+    def __len__(self) -> int:
         return len(self._data)
 
-    def __iter__(self):
+    def __iter__(self) -> Iterator[str | int | None]:
         return iter(self._data)
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: str | int | None) -> Any:
         orig = val = self._data[key]
         if isinstance(val, Alias):  # resolve an alias
             val = val.resolve(self.base)
@@ -242,17 +238,17 @@ class LocaleDataDict(abc.MutableMapping):
             alias, others = val
             val = alias.resolve(self.base).copy()
             merge(val, others)
-        if type(val) is dict:  # Return a nested alias-resolving dict
+        if isinstance(val, dict):  # Return a nested alias-resolving dict
             val = LocaleDataDict(val, base=self.base)
         if val is not orig:
             self._data[key] = val
         return val
 
-    def __setitem__(self, key, value):
+    def __setitem__(self, key: str | int | None, value: Any) -> None:
         self._data[key] = value
 
-    def __delitem__(self, key):
+    def __delitem__(self, key: str | int | None) -> None:
         del self._data[key]
 
-    def copy(self):
+    def copy(self) -> LocaleDataDict:
         return LocaleDataDict(self._data.copy(), base=self.base)
