@@ -1,25 +1,30 @@
-# -*- coding: utf-8 -*-
 """
     babel.numbers
     ~~~~~~~~~~~~~
 
     CLDR Plural support.  See UTS #35.
 
-    :copyright: (c) 2013-2022 by the Babel Team.
+    :copyright: (c) 2013-2023 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
+from __future__ import annotations
+
 import decimal
 import re
+from collections.abc import Iterable, Mapping
+from typing import TYPE_CHECKING, Any, Callable
 
+if TYPE_CHECKING:
+    from typing_extensions import Literal
 
 _plural_tags = ('zero', 'one', 'two', 'few', 'many', 'other')
 _fallback_tag = 'other'
 
 
-def extract_operands(source):
+def extract_operands(source: float | decimal.Decimal) -> tuple[decimal.Decimal | int, int, int, int, int, int, Literal[0], Literal[0]]:
     """Extract operands from a decimal, a float or an int, according to `CLDR rules`_.
 
-    The result is a 8-tuple (n, i, v, w, f, t, c, e), where those symbols are as follows:
+    The result is an 8-tuple (n, i, v, w, f, t, c, e), where those symbols are as follows:
 
     ====== ===============================================================
     Symbol Value
@@ -75,7 +80,7 @@ def extract_operands(source):
     return n, i, v, w, f, t, c, e
 
 
-class PluralRule(object):
+class PluralRule:
     """Represents a set of language pluralization rules.  The constructor
     accepts a list of (tag, expr) tuples or a dict of `CLDR rules`_. The
     resulting object is callable and accepts one parameter with a positive or
@@ -98,7 +103,7 @@ class PluralRule(object):
 
     __slots__ = ('abstract', '_func')
 
-    def __init__(self, rules):
+    def __init__(self, rules: Mapping[str, str] | Iterable[tuple[str, str]]) -> None:
         """Initialize the rule instance.
 
         :param rules: a list of ``(tag, expr)``) tuples with the rules
@@ -106,42 +111,39 @@ class PluralRule(object):
                       and expressions as values.
         :raise RuleError: if the expression is malformed
         """
-        if isinstance(rules, dict):
+        if isinstance(rules, Mapping):
             rules = rules.items()
         found = set()
-        self.abstract = []
-        for key, expr in sorted(list(rules)):
+        self.abstract: list[tuple[str, Any]] = []
+        for key, expr in sorted(rules):
             if key not in _plural_tags:
-                raise ValueError('unknown tag %r' % key)
+                raise ValueError(f"unknown tag {key!r}")
             elif key in found:
-                raise ValueError('tag %r defined twice' % key)
+                raise ValueError(f"tag {key!r} defined twice")
             found.add(key)
             ast = _Parser(expr).ast
             if ast:
                 self.abstract.append((key, ast))
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         rules = self.rules
-        return '<%s %r>' % (
-            type(self).__name__,
-            ', '.join(['%s: %s' % (tag, rules[tag]) for tag in _plural_tags
-                       if tag in rules])
-        )
+        args = ", ".join([f"{tag}: {rules[tag]}" for tag in _plural_tags if tag in rules])
+        return f"<{type(self).__name__} {args!r}>"
 
     @classmethod
-    def parse(cls, rules):
+    def parse(cls, rules: Mapping[str, str] | Iterable[tuple[str, str]] | PluralRule) -> PluralRule:
         """Create a `PluralRule` instance for the given rules.  If the rules
         are a `PluralRule` object, that object is returned.
 
         :param rules: the rules as list or dict, or a `PluralRule` object
         :raise RuleError: if the expression is malformed
         """
-        if isinstance(rules, cls):
+        if isinstance(rules, PluralRule):
             return rules
         return cls(rules)
 
     @property
-    def rules(self):
+    def rules(self) -> Mapping[str, str]:
         """The `PluralRule` as a dict of unicode plural rules.
 
         >>> rule = PluralRule({'one': 'n is 1'})
@@ -149,26 +151,29 @@ class PluralRule(object):
         {'one': 'n is 1'}
         """
         _compile = _UnicodeCompiler().compile
-        return dict([(tag, _compile(ast)) for tag, ast in self.abstract])
+        return {tag: _compile(ast) for tag, ast in self.abstract}
 
-    tags = property(lambda x: frozenset([i[0] for i in x.abstract]), doc="""
-        A set of explicitly defined tags in this rule.  The implicit default
+    @property
+    def tags(self) -> frozenset[str]:
+        """A set of explicitly defined tags in this rule.  The implicit default
         ``'other'`` rules is not part of this set unless there is an explicit
-        rule for it.""")
+        rule for it.
+        """
+        return frozenset(i[0] for i in self.abstract)
 
-    def __getstate__(self):
+    def __getstate__(self) -> list[tuple[str, Any]]:
         return self.abstract
 
-    def __setstate__(self, abstract):
+    def __setstate__(self, abstract: list[tuple[str, Any]]) -> None:
         self.abstract = abstract
 
-    def __call__(self, n):
+    def __call__(self, n: float | decimal.Decimal) -> str:
         if not hasattr(self, '_func'):
             self._func = to_python(self)
         return self._func(n)
 
 
-def to_javascript(rule):
+def to_javascript(rule: Mapping[str, str] | Iterable[tuple[str, str]] | PluralRule) -> str:
     """Convert a list/dict of rules or a `PluralRule` object into a JavaScript
     function.  This function depends on no external library:
 
@@ -186,12 +191,12 @@ def to_javascript(rule):
     to_js = _JavaScriptCompiler().compile
     result = ['(function(n) { return ']
     for tag, ast in PluralRule.parse(rule).abstract:
-        result.append('%s ? %r : ' % (to_js(ast), tag))
+        result.append(f"{to_js(ast)} ? {tag!r} : ")
     result.append('%r; })' % _fallback_tag)
     return ''.join(result)
 
 
-def to_python(rule):
+def to_python(rule: Mapping[str, str] | Iterable[tuple[str, str]] | PluralRule) -> Callable[[float | decimal.Decimal], str]:
     """Convert a list/dict of rules or a `PluralRule` object into a regular
     Python function.  This is useful in situations where you need a real
     function and don't are about the actual rule object:
@@ -224,14 +229,14 @@ def to_python(rule):
     for tag, ast in PluralRule.parse(rule).abstract:
         # the str() call is to coerce the tag to the native string.  It's
         # a limited ascii restricted set of tags anyways so that is fine.
-        result.append(' if (%s): return %r' % (to_python_func(ast), str(tag)))
-    result.append(' return %r' % _fallback_tag)
+        result.append(f" if ({to_python_func(ast)}): return {str(tag)!r}")
+    result.append(f" return {_fallback_tag!r}")
     code = compile('\n'.join(result), '<rule>', 'exec')
     eval(code, namespace)
     return namespace['evaluate']
 
 
-def to_gettext(rule):
+def to_gettext(rule: Mapping[str, str] | Iterable[tuple[str, str]] | PluralRule) -> str:
     """The plural rule as gettext expression.  The gettext expression is
     technically limited to integers and returns indices rather than tags.
 
@@ -247,14 +252,14 @@ def to_gettext(rule):
     _compile = _GettextCompiler().compile
     _get_index = [tag for tag in _plural_tags if tag in used_tags].index
 
-    result = ['nplurals=%d; plural=(' % len(used_tags)]
+    result = [f"nplurals={len(used_tags)}; plural=("]
     for tag, ast in rule.abstract:
-        result.append('%s ? %d : ' % (_compile(ast), _get_index(tag)))
-    result.append('%d);' % _get_index(_fallback_tag))
+        result.append(f"{_compile(ast)} ? {_get_index(tag)} : ")
+    result.append(f"{_get_index(_fallback_tag)});")
     return ''.join(result)
 
 
-def in_range_list(num, range_list):
+def in_range_list(num: float | decimal.Decimal, range_list: Iterable[Iterable[float | decimal.Decimal]]) -> bool:
     """Integer range list test.  This is the callback for the "in" operator
     of the UTS #35 pluralization rule language:
 
@@ -274,7 +279,7 @@ def in_range_list(num, range_list):
     return num == int(num) and within_range_list(num, range_list)
 
 
-def within_range_list(num, range_list):
+def within_range_list(num: float | decimal.Decimal, range_list: Iterable[Iterable[float | decimal.Decimal]]) -> bool:
     """Float range test.  This is the callback for the "within" operator
     of the UTS #35 pluralization rule language:
 
@@ -294,7 +299,7 @@ def within_range_list(num, range_list):
     return any(num >= min_ and num <= max_ for min_, max_ in range_list)
 
 
-def cldr_modulo(a, b):
+def cldr_modulo(a: float, b: float) -> float:
     """Javaish modulo.  This modulo operator returns the value with the sign
     of the dividend rather than the divisor like Python does:
 
@@ -320,6 +325,7 @@ def cldr_modulo(a, b):
 class RuleError(Exception):
     """Raised if a rule is malformed."""
 
+
 _VARS = {
     'n',  # absolute value of the source number.
     'i',  # integer digits of n.
@@ -328,10 +334,10 @@ _VARS = {
     'f',  # visible fraction digits in n, with trailing zeros.*
     't',  # visible fraction digits in n, without trailing zeros.*
     'c',  # compact decimal exponent value: exponent of the power of 10 used in compact decimal formatting.
-    'e',  # currently, synonym for ‘c’. however, may be redefined in the future.
+    'e',  # currently, synonym for `c`. however, may be redefined in the future.
 }
 
-_RULES = [
+_RULES: list[tuple[str | None, re.Pattern[str]]] = [
     (None, re.compile(r'\s+', re.UNICODE)),
     ('word', re.compile(fr'\b(and|or|is|(?:with)?in|not|mod|[{"".join(_VARS)}])\b')),
     ('value', re.compile(r'\d+')),
@@ -340,9 +346,9 @@ _RULES = [
 ]
 
 
-def tokenize_rule(s):
+def tokenize_rule(s: str) -> list[tuple[str, str]]:
     s = s.split('@')[0]
-    result = []
+    result: list[tuple[str, str]] = []
     pos = 0
     end = len(s)
     while pos < end:
@@ -359,33 +365,39 @@ def tokenize_rule(s):
     return result[::-1]
 
 
-def test_next_token(tokens, type_, value=None):
+def test_next_token(
+    tokens: list[tuple[str, str]],
+    type_: str,
+    value: str | None = None,
+) -> list[tuple[str, str]] | bool:
     return tokens and tokens[-1][0] == type_ and \
         (value is None or tokens[-1][1] == value)
 
 
-def skip_token(tokens, type_, value=None):
+def skip_token(tokens: list[tuple[str, str]], type_: str, value: str | None = None):
     if test_next_token(tokens, type_, value):
         return tokens.pop()
 
 
-def value_node(value):
+def value_node(value: int) -> tuple[Literal['value'], tuple[int]]:
     return 'value', (value, )
 
 
-def ident_node(name):
+def ident_node(name: str) -> tuple[str, tuple[()]]:
     return name, ()
 
 
-def range_list_node(range_list):
+def range_list_node(
+    range_list: Iterable[Iterable[float | decimal.Decimal]],
+) -> tuple[Literal['range_list'], Iterable[Iterable[float | decimal.Decimal]]]:
     return 'range_list', range_list
 
 
-def negate(rv):
+def negate(rv: tuple[Any, ...]) -> tuple[Literal['not'], tuple[tuple[Any, ...]]]:
     return 'not', (rv,)
 
 
-class _Parser(object):
+class _Parser:
     """Internal parser.  This class can translate a single rule into an abstract
     tree of tuples. It implements the following grammar::
 
@@ -415,7 +427,7 @@ class _Parser(object):
       'n in 3,5,7..15'.
     - Samples are ignored.
 
-    The translator parses the expression on instanciation into an attribute
+    The translator parses the expression on instantiation into an attribute
     called `ast`.
     """
 
@@ -428,8 +440,7 @@ class _Parser(object):
             return
         self.ast = self.condition()
         if self.tokens:
-            raise RuleError('Expected end of rule, got %r' %
-                            self.tokens[-1][1])
+            raise RuleError(f"Expected end of rule, got {self.tokens[-1][1]!r}")
 
     def expect(self, type_, value=None, term=None):
         token = skip_token(self.tokens, type_, value)
@@ -438,8 +449,8 @@ class _Parser(object):
         if term is None:
             term = repr(value is None and type_ or value)
         if not self.tokens:
-            raise RuleError('expected %s but end of rule reached' % term)
-        raise RuleError('expected %s but got %r' % (term, self.tokens[-1][1]))
+            raise RuleError(f"expected {term} but end of rule reached")
+        raise RuleError(f"expected {term} but got {self.tokens[-1][1]!r}")
 
     def condition(self):
         op = self.and_condition()
@@ -510,7 +521,7 @@ class _Parser(object):
 
 def _binary_compiler(tmpl):
     """Compiler factory for the `_Compiler`."""
-    return lambda self, l, r: tmpl % (self.compile(l), self.compile(r))
+    return lambda self, left, right: tmpl % (self.compile(left), self.compile(right))
 
 
 def _unary_compiler(tmpl):
@@ -521,14 +532,14 @@ def _unary_compiler(tmpl):
 compile_zero = lambda x: '0'
 
 
-class _Compiler(object):
+class _Compiler:
     """The compilers are able to transform the expressions into multiple
     output formats.
     """
 
     def compile(self, arg):
         op, args = arg
-        return getattr(self, 'compile_' + op)(*args)
+        return getattr(self, f"compile_{op}")(*args)
 
     compile_n = lambda x: 'n'
     compile_i = lambda x: 'i'
@@ -559,11 +570,8 @@ class _PythonCompiler(_Compiler):
     compile_mod = _binary_compiler('MOD(%s, %s)')
 
     def compile_relation(self, method, expr, range_list):
-        compile_range_list = '[%s]' % ','.join(
-            ['(%s, %s)' % tuple(map(self.compile, range_))
-             for range_ in range_list[1]])
-        return '%s(%s, %s)' % (method.upper(), self.compile(expr),
-                               compile_range_list)
+        ranges = ",".join([f"({self.compile(a)}, {self.compile(b)})" for (a, b) in range_list[1]])
+        return f"{method.upper()}({self.compile(expr)}, [{ranges}])"
 
 
 class _GettextCompiler(_Compiler):
@@ -580,19 +588,11 @@ class _GettextCompiler(_Compiler):
         expr = self.compile(expr)
         for item in range_list[1]:
             if item[0] == item[1]:
-                rv.append('(%s == %s)' % (
-                    expr,
-                    self.compile(item[0])
-                ))
+                rv.append(f"({expr} == {self.compile(item[0])})")
             else:
                 min, max = map(self.compile, item)
-                rv.append('(%s >= %s && %s <= %s)' % (
-                    expr,
-                    min,
-                    expr,
-                    max
-                ))
-        return '(%s)' % ' || '.join(rv)
+                rv.append(f"({expr} >= {min} && {expr} <= {max})")
+        return f"({' || '.join(rv)})"
 
 
 class _JavaScriptCompiler(_GettextCompiler):
@@ -611,7 +611,7 @@ class _JavaScriptCompiler(_GettextCompiler):
             self, method, expr, range_list)
         if method == 'in':
             expr = self.compile(expr)
-            code = '(parseInt(%s, 10) == %s && %s)' % (expr, expr, code)
+            code = f"(parseInt({expr}, 10) == {expr} && {code})"
         return code
 
 
@@ -629,7 +629,7 @@ class _UnicodeCompiler(_Compiler):
     compile_mod = _binary_compiler('%s mod %s')
 
     def compile_not(self, relation):
-        return self.compile_relation(negated=True, *relation[1])
+        return self.compile_relation(*relation[1], negated=True)
 
     def compile_relation(self, method, expr, range_list, negated=False):
         ranges = []
@@ -637,8 +637,5 @@ class _UnicodeCompiler(_Compiler):
             if item[0] == item[1]:
                 ranges.append(self.compile(item[0]))
             else:
-                ranges.append('%s..%s' % tuple(map(self.compile, item)))
-        return '%s%s %s %s' % (
-            self.compile(expr), negated and ' not' or '',
-            method, ','.join(ranges)
-        )
+                ranges.append(f"{self.compile(item[0])}..{self.compile(item[1])}")
+        return f"{self.compile(expr)}{' not' if negated else ''} {method} {','.join(ranges)}"
