@@ -5,4 +5,4 @@ from gettext import gettext as _
 def foo():
     # TRANSLATOR: This will be a translator coment,
     # that will include several lines
-    print _('bar')
+    print(_('bar'))
