@@ -6,4 +6,4 @@ from gettext import ngettext
 def foo():
     # Note: This will have the TRANSLATOR: tag but shouldn't
     # be included on the extracted stuff
-    print ngettext('foobar', 'foobars', 1)
+    print(ngettext('foobar', 'foobars', 1))
