@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2007-2011 Edgewall Software, 2013-2022 the Babel team
+# Copyright (C) 2007-2011 Edgewall Software, 2013-2023 the Babel team
 # All rights reserved.
 #
 # This software is licensed as described in the file LICENSE, which
@@ -15,7 +14,7 @@ import os
 import unittest
 from io import BytesIO
 
-from babel.messages import mofile, Catalog
+from babel.messages import Catalog, mofile
 from babel.support import Translations
 
 
@@ -29,12 +28,11 @@ class ReadMoTestCase(unittest.TestCase):
                                'LC_MESSAGES', 'messages.mo')
         with open(mo_path, 'rb') as mo_file:
             catalog = mofile.read_mo(mo_file)
-            self.assertEqual(2, len(catalog))
-            self.assertEqual('TestProject', catalog.project)
-            self.assertEqual('0.1', catalog.version)
-            self.assertEqual('Stange', catalog['bar'].string)
-            self.assertEqual(['Fuhstange', 'Fuhstangen'],
-                             catalog['foobar'].string)
+            assert len(catalog) == 2
+            assert catalog.project == 'TestProject'
+            assert catalog.version == '0.1'
+            assert catalog['bar'].string == 'Stange'
+            assert catalog['foobar'].string == ['Fuhstange', 'Fuhstangen']
 
 
 class WriteMoTestCase(unittest.TestCase):
@@ -44,27 +42,22 @@ class WriteMoTestCase(unittest.TestCase):
         # can be applied to all subsequent messages by GNUTranslations
         # (ensuring all messages are safely converted to unicode)
         catalog = Catalog(locale='en_US')
-        catalog.add(u'', '''\
+        catalog.add('', '''\
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n''')
-        catalog.add(u'foo', 'Voh')
-        catalog.add((u'There is', u'There are'), (u'Es gibt', u'Es gibt'))
-        catalog.add(u'Fizz', '')
+        catalog.add('foo', 'Voh')
+        catalog.add(('There is', 'There are'), ('Es gibt', 'Es gibt'))
+        catalog.add('Fizz', '')
         catalog.add(('Fuzz', 'Fuzzes'), ('', ''))
         buf = BytesIO()
         mofile.write_mo(buf, catalog)
         buf.seek(0)
         translations = Translations(fp=buf)
-        self.assertEqual(u'Voh', translations.ugettext('foo'))
-        assert isinstance(translations.ugettext('foo'), str)
-        self.assertEqual(u'Es gibt', translations.ungettext('There is', 'There are', 1))
-        assert isinstance(translations.ungettext('There is', 'There are', 1), str)
-        self.assertEqual(u'Fizz', translations.ugettext('Fizz'))
-        assert isinstance(translations.ugettext('Fizz'), str)
-        self.assertEqual(u'Fuzz', translations.ugettext('Fuzz'))
-        assert isinstance(translations.ugettext('Fuzz'), str)
-        self.assertEqual(u'Fuzzes', translations.ugettext('Fuzzes'))
-        assert isinstance(translations.ugettext('Fuzzes'), str)
+        assert translations.ugettext('foo') == 'Voh'
+        assert translations.ungettext('There is', 'There are', 1) == 'Es gibt'
+        assert translations.ugettext('Fizz') == 'Fizz'
+        assert translations.ugettext('Fuzz') == 'Fuzz'
+        assert translations.ugettext('Fuzzes') == 'Fuzzes'
 
     def test_more_plural_forms(self):
         catalog2 = Catalog(locale='ru_RU')
@@ -74,18 +67,18 @@ class WriteMoTestCase(unittest.TestCase):
 
     def test_empty_translation_with_fallback(self):
         catalog1 = Catalog(locale='fr_FR')
-        catalog1.add(u'', '''\
+        catalog1.add('', '''\
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n''')
-        catalog1.add(u'Fuzz', '')
+        catalog1.add('Fuzz', '')
         buf1 = BytesIO()
         mofile.write_mo(buf1, catalog1)
         buf1.seek(0)
         catalog2 = Catalog(locale='fr')
-        catalog2.add(u'', '''\
+        catalog2.add('', '''\
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n''')
-        catalog2.add(u'Fuzz', 'Flou')
+        catalog2.add('Fuzz', 'Flou')
         buf2 = BytesIO()
         mofile.write_mo(buf2, catalog2)
         buf2.seek(0)
@@ -93,4 +86,4 @@ class WriteMoTestCase(unittest.TestCase):
         translations = Translations(fp=buf1)
         translations.add_fallback(Translations(fp=buf2))
 
-        self.assertEqual(u'Flou', translations.ugettext('Fuzz'))
+        assert translations.ugettext('Fuzz') == 'Flou'
