@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2007-2011 Edgewall Software, 2013-2022 the Babel team
+# Copyright (C) 2007-2011 Edgewall Software, 2013-2023 the Babel team
 # All rights reserved.
 #
 # This software is licensed as described in the file LICENSE, which
@@ -11,22 +10,23 @@
 # individuals. For the exact contribution history, see the revision
 # history and logs, available at http://babel.edgewall.org/log/.
 
+import datetime
 import inspect
 import os
 import shutil
+import sys
 import tempfile
 import unittest
-import pytest
-import sys
-from datetime import date, datetime, timedelta
 from io import BytesIO
+
+import pytest
 
 from babel import support
 from babel.messages import Catalog
 from babel.messages.mofile import write_mo
 
-
 SKIP_LGETTEXT = sys.version_info >= (3, 8)
+
 
 @pytest.mark.usefixtures("os_environ")
 class TranslationsTestCase(unittest.TestCase):
@@ -63,9 +63,8 @@ class TranslationsTestCase(unittest.TestCase):
         self.translations = translations1.add(translations2, merge=False)
 
     def assertEqualTypeToo(self, expected, result):
-        self.assertEqual(expected, result)
-        assert type(expected) == type(result), "instance type's do not " + \
-            "match: %r!=%r" % (type(expected), type(result))
+        assert expected == result
+        assert type(expected) == type(result), f"instance types do not match: {type(expected)!r}!={type(result)!r}"
 
     def test_pgettext(self):
         self.assertEqualTypeToo('Voh', self.translations.gettext('foo'))
@@ -73,9 +72,9 @@ class TranslationsTestCase(unittest.TestCase):
                                                                      'foo'))
 
     def test_upgettext(self):
-        self.assertEqualTypeToo(u'Voh', self.translations.ugettext('foo'))
-        self.assertEqualTypeToo(u'VohCTX', self.translations.upgettext('foo',
-                                                                       'foo'))
+        self.assertEqualTypeToo('Voh', self.translations.ugettext('foo'))
+        self.assertEqualTypeToo('VohCTX', self.translations.upgettext('foo',
+                                                                      'foo'))
 
     @pytest.mark.skipif(SKIP_LGETTEXT, reason="lgettext is deprecated")
     def test_lpgettext(self):
@@ -96,14 +95,14 @@ class TranslationsTestCase(unittest.TestCase):
                                                             'foos1', 2))
 
     def test_unpgettext(self):
-        self.assertEqualTypeToo(u'Voh1',
+        self.assertEqualTypeToo('Voh1',
                                 self.translations.ungettext('foo1', 'foos1', 1))
-        self.assertEqualTypeToo(u'Vohs1',
+        self.assertEqualTypeToo('Vohs1',
                                 self.translations.ungettext('foo1', 'foos1', 2))
-        self.assertEqualTypeToo(u'VohCTX1',
+        self.assertEqualTypeToo('VohCTX1',
                                 self.translations.unpgettext('foo', 'foo1',
                                                              'foos1', 1))
-        self.assertEqualTypeToo(u'VohsCTX1',
+        self.assertEqualTypeToo('VohsCTX1',
                                 self.translations.unpgettext('foo', 'foo1',
                                                              'foos1', 2))
 
@@ -128,9 +127,9 @@ class TranslationsTestCase(unittest.TestCase):
 
     def test_dupgettext(self):
         self.assertEqualTypeToo(
-            u'VohD', self.translations.dugettext('messages1', 'foo'))
+            'VohD', self.translations.dugettext('messages1', 'foo'))
         self.assertEqualTypeToo(
-            u'VohCTXD', self.translations.dupgettext('messages1', 'foo', 'foo'))
+            'VohCTXD', self.translations.dupgettext('messages1', 'foo', 'foo'))
 
     @pytest.mark.skipif(SKIP_LGETTEXT, reason="lgettext is deprecated")
     def test_ldpgettext(self):
@@ -153,15 +152,15 @@ class TranslationsTestCase(unittest.TestCase):
 
     def test_dunpgettext(self):
         self.assertEqualTypeToo(
-            u'VohD1', self.translations.dungettext('messages1', 'foo1', 'foos1', 1))
+            'VohD1', self.translations.dungettext('messages1', 'foo1', 'foos1', 1))
         self.assertEqualTypeToo(
-            u'VohsD1', self.translations.dungettext('messages1', 'foo1', 'foos1', 2))
+            'VohsD1', self.translations.dungettext('messages1', 'foo1', 'foos1', 2))
         self.assertEqualTypeToo(
-            u'VohCTXD1', self.translations.dunpgettext('messages1', 'foo', 'foo1',
-                                                       'foos1', 1))
+            'VohCTXD1', self.translations.dunpgettext('messages1', 'foo', 'foo1',
+                                                      'foos1', 1))
         self.assertEqualTypeToo(
-            u'VohsCTXD1', self.translations.dunpgettext('messages1', 'foo', 'foo1',
-                                                        'foos1', 2))
+            'VohsCTXD1', self.translations.dunpgettext('messages1', 'foo', 'foo1',
+                                                       'foos1', 2))
 
     @pytest.mark.skipif(SKIP_LGETTEXT, reason="lgettext is deprecated")
     def test_ldnpgettext(self):
@@ -187,7 +186,7 @@ class TranslationsTestCase(unittest.TestCase):
                 write_mo(f, catalog)
 
             translations = support.Translations.load(tempdir, locales=('fr',), domain='messages')
-            self.assertEqual('bar', translations.gettext('foo'))
+            assert translations.gettext('foo') == 'bar'
         finally:
             shutil.rmtree(tempdir)
 
@@ -211,22 +210,19 @@ class NullTranslationsTestCase(unittest.TestCase):
     def test_same_methods(self):
         for name in self.method_names():
             if not hasattr(self.null_translations, name):
-                self.fail('NullTranslations does not provide method %r' % name)
+                self.fail(f"NullTranslations does not provide method {name!r}")
 
     def test_method_signature_compatibility(self):
         for name in self.method_names():
             translations_method = getattr(self.translations, name)
             null_method = getattr(self.null_translations, name)
-            self.assertEqual(
-                inspect.getfullargspec(translations_method),
-                inspect.getfullargspec(null_method),
-            )
+            assert inspect.getfullargspec(translations_method) == inspect.getfullargspec(null_method)
 
     def test_same_return_values(self):
         data = {
-            'message': u'foo', 'domain': u'domain', 'context': 'tests',
-            'singular': u'bar', 'plural': u'baz', 'num': 1,
-            'msgid1': u'bar', 'msgid2': u'baz', 'n': 1,
+            'message': 'foo', 'domain': 'domain', 'context': 'tests',
+            'singular': 'bar', 'plural': 'baz', 'num': 1,
+            'msgid1': 'bar', 'msgid2': 'baz', 'n': 1,
         }
         for name in self.method_names():
             method = getattr(self.translations, name)
@@ -234,7 +230,7 @@ class NullTranslationsTestCase(unittest.TestCase):
             signature = inspect.getfullargspec(method)
             parameter_names = [name for name in signature.args if name != 'self']
             values = [data[name] for name in parameter_names]
-            self.assertEqual(method(*values), null_method(*values))
+            assert method(*values) == null_method(*values)
 
 
 class LazyProxyTestCase(unittest.TestCase):
@@ -246,8 +242,8 @@ class LazyProxyTestCase(unittest.TestCase):
             self.counter += 1
             return self.counter
         proxy = support.LazyProxy(add_one)
-        self.assertEqual(1, proxy.value)
-        self.assertEqual(1, proxy.value)
+        assert proxy.value == 1
+        assert proxy.value == 1
 
     def test_can_disable_proxy_cache(self):
         self.counter = 0
@@ -256,8 +252,8 @@ class LazyProxyTestCase(unittest.TestCase):
             self.counter += 1
             return self.counter
         proxy = support.LazyProxy(add_one, enable_cache=False)
-        self.assertEqual(1, proxy.value)
-        self.assertEqual(2, proxy.value)
+        assert proxy.value == 1
+        assert proxy.value == 2
 
     def test_can_copy_proxy(self):
         from copy import copy
@@ -271,8 +267,8 @@ class LazyProxyTestCase(unittest.TestCase):
         proxy_copy = copy(proxy)
 
         numbers.pop(0)
-        self.assertEqual(2, proxy.value)
-        self.assertEqual(2, proxy_copy.value)
+        assert proxy.value == 2
+        assert proxy_copy.value == 2
 
     def test_can_deepcopy_proxy(self):
         from copy import deepcopy
@@ -285,8 +281,8 @@ class LazyProxyTestCase(unittest.TestCase):
         proxy_deepcopy = deepcopy(proxy)
 
         numbers.pop(0)
-        self.assertEqual(2, proxy.value)
-        self.assertEqual(1, proxy_deepcopy.value)
+        assert proxy.value == 2
+        assert proxy_deepcopy.value == 1
 
     def test_handle_attribute_error(self):
 
@@ -295,67 +291,42 @@ class LazyProxyTestCase(unittest.TestCase):
 
         proxy = support.LazyProxy(raise_attribute_error)
         with pytest.raises(AttributeError) as exception:
-            proxy.value
+            _ = proxy.value
 
-        self.assertEqual('message', str(exception.value))
-
-
-def test_format_date():
-    fmt = support.Format('en_US')
-    assert fmt.date(date(2007, 4, 1)) == 'Apr 1, 2007'
+        assert str(exception.value) == 'message'
 
 
-def test_format_datetime():
-    from pytz import timezone
-    fmt = support.Format('en_US', tzinfo=timezone('US/Eastern'))
-    when = datetime(2007, 4, 1, 15, 30)
-    assert fmt.datetime(when) == 'Apr 1, 2007, 11:30:00 AM'
+WHEN = datetime.datetime(2007, 4, 1, 15, 30)
+
+def test_format_datetime(timezone_getter):
+    fmt = support.Format('en_US', tzinfo=timezone_getter('US/Eastern'))
+    assert fmt.datetime(WHEN) == 'Apr 1, 2007, 11:30:00\u202fAM'
 
 
-def test_format_time():
-    from pytz import timezone
-    fmt = support.Format('en_US', tzinfo=timezone('US/Eastern'))
-    assert fmt.time(datetime(2007, 4, 1, 15, 30)) == '11:30:00 AM'
-
-
-def test_format_timedelta():
-    fmt = support.Format('en_US')
-    assert fmt.timedelta(timedelta(weeks=11)) == '3 months'
-
-
-def test_format_number():
-    fmt = support.Format('en_US')
-    assert fmt.number(1099) == '1,099'
-
-
-def test_format_decimal():
-    fmt = support.Format('en_US')
-    assert fmt.decimal(1.2345) == '1.234'
-
-
-def test_format_percent():
-    fmt = support.Format('en_US')
-    assert fmt.percent(0.34) == '34%'
+def test_format_time(timezone_getter):
+    fmt = support.Format('en_US', tzinfo=timezone_getter('US/Eastern'))
+    assert fmt.time(WHEN) == '11:30:00\u202fAM'
 
 
 def test_lazy_proxy():
     def greeting(name='world'):
-        return u'Hello, %s!' % name
-    lazy_greeting = support.LazyProxy(greeting, name='Joe')
-    assert str(lazy_greeting) == u"Hello, Joe!"
-    assert u'  ' + lazy_greeting == u'  Hello, Joe!'
-    assert u'(%s)' % lazy_greeting == u'(Hello, Joe!)'
+        return f"Hello, {name}!"
 
-    greetings = [
+    lazy_greeting = support.LazyProxy(greeting, name='Joe')
+    assert str(lazy_greeting) == "Hello, Joe!"
+    assert '  ' + lazy_greeting == '  Hello, Joe!'
+    assert '(%s)' % lazy_greeting == '(Hello, Joe!)'
+    assert f"[{lazy_greeting}]" == "[Hello, Joe!]"
+
+    greetings = sorted([
         support.LazyProxy(greeting, 'world'),
         support.LazyProxy(greeting, 'Joe'),
         support.LazyProxy(greeting, 'universe'),
-    ]
-    greetings.sort()
+    ])
     assert [str(g) for g in greetings] == [
-        u"Hello, Joe!",
-        u"Hello, universe!",
-        u"Hello, world!",
+        "Hello, Joe!",
+        "Hello, universe!",
+        "Hello, world!",
     ]
 
 
