@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """
     babel
     ~~~~~
@@ -13,12 +12,26 @@
        access to various locale display names, localized number and date
        formatting, etc.
 
-    :copyright: (c) 2013-2022 by the Babel Team.
+    :copyright: (c) 2013-2023 by the Babel Team.
     :license: BSD, see LICENSE for more details.
 """
 
-from babel.core import UnknownLocaleError, Locale, default_locale, \
-    negotiate_locale, parse_locale, get_locale_identifier
+from babel.core import (
+    Locale,
+    UnknownLocaleError,
+    default_locale,
+    get_locale_identifier,
+    negotiate_locale,
+    parse_locale,
+)
 
+__version__ = '2.13.1'
 
-__version__ = '2.10.1'
+__all__ = [
+    'Locale',
+    'UnknownLocaleError',
+    'default_locale',
+    'get_locale_identifier',
+    'negotiate_locale',
+    'parse_locale',
+]
