@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # Babel documentation build configuration file, created by
 # sphinx-quickstart on Wed Jul  3 17:53:01 2013.
@@ -11,7 +10,8 @@
 # All configuration values have a default; values that are commented out
 # serve to show the default.
 
-import sys, os
+import os
+import sys
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -43,17 +43,17 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'Babel'
-copyright = u'2022, The Babel Team'
+project = 'Babel'
+copyright = '2023, The Babel Team'
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 #
 # The short X.Y version.
-version = '2.10'
+version = '2.13'
 # The full version, including alpha/beta/rc tags.
-release = '2.10.1'
+release = '2.13.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
@@ -194,8 +194,8 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-  ('index', 'Babel.tex', u'Babel Documentation',
-   u'The Babel Team', 'manual'),
+  ('index', 'Babel.tex', 'Babel Documentation',
+   'The Babel Team', 'manual'),
 ]
 
 # The name of an image file (relative to this directory) to place at the top of
@@ -224,7 +224,7 @@ latex_logo = '_static/logo.png'
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', 'babel', u'Babel Documentation', [u'The Babel Team'], 1),
+    ('index', 'babel', 'Babel Documentation', ['The Babel Team'], 1),
 ]
 
 # If true, show URL addresses after external links.
@@ -237,8 +237,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index_', 'Babel', u'Babel Documentation',
-   u'The Babel Team', 'Babel', 'One line description of project.',
+  ('index_', 'Babel', 'Babel Documentation',
+   'The Babel Team', 'Babel', 'One line description of project.',
    'Miscellaneous'),
 ]
 
@@ -252,10 +252,10 @@ texinfo_documents = [
 #texinfo_show_urls = 'footnote'
 
 intersphinx_mapping = {
-    'https://docs.python.org/2/': None,
+    'https://docs.python.org/3/': None,
 }
 
 extlinks = {
-    'gh': ('https://github.com/python-babel/babel/issues/%s', '#'),
-    'trac': ('http://babel.edgewall.org/ticket/%s', 'ticket #'),
+    'gh': ('https://github.com/python-babel/babel/issues/%s', '#%s'),
+    'trac': ('http://babel.edgewall.org/ticket/%s', 'ticket #%s'),
 }
