@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2007-2011 Edgewall Software, 2013-2022 the Babel team
+# Copyright (C) 2007-2011 Edgewall Software, 2013-2023 the Babel team
 # All rights reserved.
 #
 # This software is licensed as described in the file LICENSE, which
@@ -15,6 +14,8 @@ import codecs
 import sys
 import unittest
 from io import BytesIO, StringIO
+
+import pytest
 
 from babel.messages import extract
 
@@ -37,18 +38,18 @@ msg10 = dngettext(getDomain(), 'Page', 'Pages', 3)
         messages = list(extract.extract_python(buf,
                                                extract.DEFAULT_KEYWORDS.keys(),
                                                [], {}))
-        self.assertEqual([
+        assert messages == [
             (1, '_', None, []),
             (2, 'ungettext', (None, None, None), []),
-            (3, 'ungettext', (u'Babel', None, None), []),
-            (4, 'ungettext', (None, u'Babels', None), []),
-            (5, 'ungettext', (u'bunny', u'bunnies', None), []),
-            (6, 'ungettext', (None, u'bunnies', None), []),
+            (3, 'ungettext', ('Babel', None, None), []),
+            (4, 'ungettext', (None, 'Babels', None), []),
+            (5, 'ungettext', ('bunny', 'bunnies', None), []),
+            (6, 'ungettext', (None, 'bunnies', None), []),
             (7, '_', None, []),
-            (8, 'gettext', u'Rabbit', []),
-            (9, 'dgettext', (u'wiki', None), []),
-            (10, 'dngettext', (None, u'Page', u'Pages', None), [])],
-            messages)
+            (8, 'gettext', 'Rabbit', []),
+            (9, 'dgettext', ('wiki', None), []),
+            (10, 'dngettext', (None, 'Page', 'Pages', None), []),
+        ]
 
     def test_extract_default_encoding_ascii(self):
         buf = BytesIO(b'_("a")')
@@ -56,14 +57,14 @@ msg10 = dngettext(getDomain(), 'Page', 'Pages', 3)
             buf, list(extract.DEFAULT_KEYWORDS), [], {},
         ))
         # Should work great in both py2 and py3
-        self.assertEqual([(1, '_', 'a', [])], messages)
+        assert messages == [(1, '_', 'a', [])]
 
     def test_extract_default_encoding_utf8(self):
-        buf = BytesIO(u'_("☃")'.encode('UTF-8'))
+        buf = BytesIO('_("☃")'.encode('UTF-8'))
         messages = list(extract.extract_python(
             buf, list(extract.DEFAULT_KEYWORDS), [], {},
         ))
-        self.assertEqual([(1, '_', u'☃', [])], messages)
+        assert messages == [(1, '_', '☃', [])]
 
     def test_nested_comments(self):
         buf = BytesIO(b"""\
@@ -73,8 +74,7 @@ msg = ngettext('pylon',  # TRANSLATORS: shouldn't be
 """)
         messages = list(extract.extract_python(buf, ('ngettext',),
                                                ['TRANSLATORS:'], {}))
-        self.assertEqual([(1, 'ngettext', (u'pylon', u'pylons', None), [])],
-                         messages)
+        assert messages == [(1, 'ngettext', ('pylon', 'pylons', None), [])]
 
     def test_comments_with_calls_that_spawn_multiple_lines(self):
         buf = BytesIO(b"""\
@@ -97,23 +97,10 @@ add_notice(req, ngettext("Bar deleted.",
         messages = list(extract.extract_python(buf, ('ngettext', '_'), ['NOTE:'],
 
                                                {'strip_comment_tags': False}))
-        self.assertEqual((6, '_', 'Locale deleted.',
-                          [u'NOTE: This Comment SHOULD Be Extracted']),
-                         messages[1])
-        self.assertEqual((10, 'ngettext', (u'Foo deleted.', u'Foos deleted.',
-                                           None),
-                          [u'NOTE: This Comment SHOULD Be Extracted']),
-                         messages[2])
-        self.assertEqual((3, 'ngettext',
-                          (u'Catalog deleted.',
-                           u'Catalogs deleted.', None),
-                          [u'NOTE: This Comment SHOULD Be Extracted']),
-                         messages[0])
-        self.assertEqual((15, 'ngettext', (u'Bar deleted.', u'Bars deleted.',
-                                           None),
-                          [u'NOTE: This Comment SHOULD Be Extracted',
-                           u'NOTE: And This One Too']),
-                         messages[3])
+        assert messages[0] == (3, 'ngettext', ('Catalog deleted.', 'Catalogs deleted.', None), ['NOTE: This Comment SHOULD Be Extracted'])
+        assert messages[1] == (6, '_', 'Locale deleted.', ['NOTE: This Comment SHOULD Be Extracted'])
+        assert messages[2] == (10, 'ngettext', ('Foo deleted.', 'Foos deleted.', None), ['NOTE: This Comment SHOULD Be Extracted'])
+        assert messages[3] == (15, 'ngettext', ('Bar deleted.', 'Bars deleted.', None), ['NOTE: This Comment SHOULD Be Extracted', 'NOTE: And This One Too'])
 
     def test_declarations(self):
         buf = BytesIO(b"""\
@@ -129,10 +116,11 @@ class Meta:
         messages = list(extract.extract_python(buf,
                                                extract.DEFAULT_KEYWORDS.keys(),
                                                [], {}))
-        self.assertEqual([(3, '_', u'Page arg 1', []),
-                          (3, '_', u'Page arg 2', []),
-                          (8, '_', u'log entry', [])],
-                         messages)
+        assert messages == [
+            (3, '_', 'Page arg 1', []),
+            (3, '_', 'Page arg 2', []),
+            (8, '_', 'log entry', [])
+        ]
 
     def test_multiline(self):
         buf = BytesIO(b"""\
@@ -143,9 +131,10 @@ msg2 = ngettext('elvis',
                  count)
 """)
         messages = list(extract.extract_python(buf, ('ngettext',), [], {}))
-        self.assertEqual([(1, 'ngettext', (u'pylon', u'pylons', None), []),
-                          (3, 'ngettext', (u'elvis', u'elvises', None), [])],
-                         messages)
+        assert messages == [
+            (1, 'ngettext', ('pylon', 'pylons', None), []),
+            (3, 'ngettext', ('elvis', 'elvises', None), []),
+        ]
 
     def test_npgettext(self):
         buf = BytesIO(b"""\
@@ -156,9 +145,10 @@ msg2 = npgettext('Strings','elvis',
                  count)
 """)
         messages = list(extract.extract_python(buf, ('npgettext',), [], {}))
-        self.assertEqual([(1, 'npgettext', (u'Strings', u'pylon', u'pylons', None), []),
-                          (3, 'npgettext', (u'Strings', u'elvis', u'elvises', None), [])],
-                         messages)
+        assert messages == [
+            (1, 'npgettext', ('Strings', 'pylon', 'pylons', None), []),
+            (3, 'npgettext', ('Strings', 'elvis', 'elvises', None), []),
+        ]
         buf = BytesIO(b"""\
 msg = npgettext('Strings', 'pylon',  # TRANSLATORS: shouldn't be
                 'pylons', # TRANSLATORS: seeing this
@@ -166,8 +156,9 @@ msg = npgettext('Strings', 'pylon',  # TRANSLATORS: shouldn't be
 """)
         messages = list(extract.extract_python(buf, ('npgettext',),
                                                ['TRANSLATORS:'], {}))
-        self.assertEqual([(1, 'npgettext', (u'Strings', u'pylon', u'pylons', None), [])],
-                         messages)
+        assert messages == [
+            (1, 'npgettext', ('Strings', 'pylon', 'pylons', None), []),
+        ]
 
     def test_triple_quoted_strings(self):
         buf = BytesIO(b"""\
@@ -178,10 +169,11 @@ msg2 = ngettext(\"\"\"elvis\"\"\", 'elvises', count)
         messages = list(extract.extract_python(buf,
                                                extract.DEFAULT_KEYWORDS.keys(),
                                                [], {}))
-        self.assertEqual([(1, '_', u'pylons', []),
-                          (2, 'ngettext', (u'elvis', u'elvises', None), []),
-                          (3, 'ngettext', (u'elvis', u'elvises', None), [])],
-                         messages)
+        assert messages == [
+            (1, '_', 'pylons', []),
+            (2, 'ngettext', ('elvis', 'elvises', None), []),
+            (3, 'ngettext', ('elvis', 'elvises', None), []),
+        ]
 
     def test_multiline_strings(self):
         buf = BytesIO(b"""\
@@ -192,12 +184,12 @@ gettext message catalog library.''')
         messages = list(extract.extract_python(buf,
                                                extract.DEFAULT_KEYWORDS.keys(),
                                                [], {}))
-        self.assertEqual(
-            [(1, '_',
-              u'This module provides internationalization and localization\n'
-              'support for your Python programs by providing an interface to '
-              'the GNU\ngettext message catalog library.', [])],
-            messages)
+        assert messages == [
+            (1, '_',
+            'This module provides internationalization and localization\n'
+            'support for your Python programs by providing an interface to '
+            'the GNU\ngettext message catalog library.', []),
+        ]
 
     def test_concatenated_strings(self):
         buf = BytesIO(b"""\
@@ -206,12 +198,12 @@ foobar = _('foo' 'bar')
         messages = list(extract.extract_python(buf,
                                                extract.DEFAULT_KEYWORDS.keys(),
                                                [], {}))
-        self.assertEqual(u'foobar', messages[0][2])
+        assert messages[0][2] == 'foobar'
 
     def test_unicode_string_arg(self):
         buf = BytesIO(b"msg = _(u'Foo Bar')")
         messages = list(extract.extract_python(buf, ('_',), [], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
+        assert messages[0][2] == 'Foo Bar'
 
     def test_comment_tag(self):
         buf = BytesIO(b"""
@@ -219,8 +211,8 @@ foobar = _('foo' 'bar')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u'NOTE: A translation comment'], messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == ['NOTE: A translation comment']
 
     def test_comment_tag_multiline(self):
         buf = BytesIO(b"""
@@ -229,9 +221,8 @@ msg = _(u'Foo Bar')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u'NOTE: A translation comment', u'with a second line'],
-                         messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == ['NOTE: A translation comment', 'with a second line']
 
     def test_translator_comments_with_previous_non_translator_comments(self):
         buf = BytesIO(b"""
@@ -242,9 +233,8 @@ msg = _(u'Foo Bar')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u'NOTE: A translation comment', u'with a second line'],
-                         messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == ['NOTE: A translation comment', 'with a second line']
 
     def test_comment_tags_not_on_start_of_comment(self):
         buf = BytesIO(b"""
@@ -255,8 +245,8 @@ msg = _(u'Foo Bar')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u'NOTE: This one will be'], messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == ['NOTE: This one will be']
 
     def test_multiple_comment_tags(self):
         buf = BytesIO(b"""
@@ -269,11 +259,10 @@ msg = _(u'Foo Bar2')
 """)
         messages = list(extract.extract_python(buf, ('_',),
                                                ['NOTE1:', 'NOTE2:'], {}))
-        self.assertEqual(u'Foo Bar1', messages[0][2])
-        self.assertEqual([u'NOTE1: A translation comment for tag1',
-                          u'with a second line'], messages[0][3])
-        self.assertEqual(u'Foo Bar2', messages[1][2])
-        self.assertEqual([u'NOTE2: A translation comment for tag2'], messages[1][3])
+        assert messages[0][2] == 'Foo Bar1'
+        assert messages[0][3] == ['NOTE1: A translation comment for tag1', 'with a second line']
+        assert messages[1][2] == 'Foo Bar2'
+        assert messages[1][3] == ['NOTE2: A translation comment for tag2']
 
     def test_two_succeeding_comments(self):
         buf = BytesIO(b"""
@@ -282,8 +271,8 @@ msg = _(u'Foo Bar2')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u'NOTE: one', u'NOTE: two'], messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == ['NOTE: one', 'NOTE: two']
 
     def test_invalid_translator_comments(self):
         buf = BytesIO(b"""
@@ -293,8 +282,8 @@ hello = 'there'
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([], messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == []
 
     def test_invalid_translator_comments2(self):
         buf = BytesIO(b"""
@@ -308,10 +297,10 @@ rows = [[v for v in range(0,10)] for row in range(0,10)]
 hello = _('Hello')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Hi there!', messages[0][2])
-        self.assertEqual([u'NOTE: Hi!'], messages[0][3])
-        self.assertEqual(u'Hello', messages[1][2])
-        self.assertEqual([], messages[1][3])
+        assert messages[0][2] == 'Hi there!'
+        assert messages[0][3] == ['NOTE: Hi!']
+        assert messages[1][2] == 'Hello'
+        assert messages[1][3] == []
 
     def test_invalid_translator_comments3(self):
         buf = BytesIO(b"""
@@ -321,8 +310,8 @@ hello = _('Hello')
 hithere = _('Hi there!')
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Hi there!', messages[0][2])
-        self.assertEqual([], messages[0][3])
+        assert messages[0][2] == 'Hi there!'
+        assert messages[0][3] == []
 
     def test_comment_tag_with_leading_space(self):
         buf = BytesIO(b"""
@@ -331,9 +320,8 @@ hithere = _('Hi there!')
 msg = _(u'Foo Bar')
 """)
         messages = list(extract.extract_python(buf, ('_',), [':'], {}))
-        self.assertEqual(u'Foo Bar', messages[0][2])
-        self.assertEqual([u': A translation comment', u': with leading spaces'],
-                         messages[0][3])
+        assert messages[0][2] == 'Foo Bar'
+        assert messages[0][3] == [': A translation comment', ': with leading spaces']
 
     def test_different_signatures(self):
         buf = BytesIO(b"""
@@ -345,66 +333,66 @@ n = ngettext()
 n = ngettext('foo')
 """)
         messages = list(extract.extract_python(buf, ('_', 'ngettext'), [], {}))
-        self.assertEqual((u'foo', u'bar'), messages[0][2])
-        self.assertEqual((u'hello', u'there', None), messages[1][2])
-        self.assertEqual((None, u'hello', u'there'), messages[2][2])
-        self.assertEqual((None, None), messages[3][2])
-        self.assertEqual(None, messages[4][2])
-        self.assertEqual('foo', messages[5][2])
+        assert messages[0][2] == ('foo', 'bar')
+        assert messages[1][2] == ('hello', 'there', None)
+        assert messages[2][2] == (None, 'hello', 'there')
+        assert messages[3][2] == (None, None)
+        assert messages[4][2] is None
+        assert messages[5][2] == 'foo'
 
     def test_utf8_message(self):
-        buf = BytesIO(u"""
+        buf = BytesIO("""
 # NOTE: hello
 msg = _('Bonjour à tous')
 """.encode('utf-8'))
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'],
                                                {'encoding': 'utf-8'}))
-        self.assertEqual(u'Bonjour à tous', messages[0][2])
-        self.assertEqual([u'NOTE: hello'], messages[0][3])
+        assert messages[0][2] == 'Bonjour à tous'
+        assert messages[0][3] == ['NOTE: hello']
 
     def test_utf8_message_with_magic_comment(self):
-        buf = BytesIO(u"""# -*- coding: utf-8 -*-
+        buf = BytesIO("""# -*- coding: utf-8 -*-
 # NOTE: hello
 msg = _('Bonjour à tous')
 """.encode('utf-8'))
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Bonjour à tous', messages[0][2])
-        self.assertEqual([u'NOTE: hello'], messages[0][3])
+        assert messages[0][2] == 'Bonjour à tous'
+        assert messages[0][3] == ['NOTE: hello']
 
     def test_utf8_message_with_utf8_bom(self):
-        buf = BytesIO(codecs.BOM_UTF8 + u"""
+        buf = BytesIO(codecs.BOM_UTF8 + """
 # NOTE: hello
 msg = _('Bonjour à tous')
 """.encode('utf-8'))
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Bonjour à tous', messages[0][2])
-        self.assertEqual([u'NOTE: hello'], messages[0][3])
+        assert messages[0][2] == 'Bonjour à tous'
+        assert messages[0][3] == ['NOTE: hello']
 
     def test_utf8_message_with_utf8_bom_and_magic_comment(self):
-        buf = BytesIO(codecs.BOM_UTF8 + u"""# -*- coding: utf-8 -*-
+        buf = BytesIO(codecs.BOM_UTF8 + """# -*- coding: utf-8 -*-
 # NOTE: hello
 msg = _('Bonjour à tous')
 """.encode('utf-8'))
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Bonjour à tous', messages[0][2])
-        self.assertEqual([u'NOTE: hello'], messages[0][3])
+        assert messages[0][2] == 'Bonjour à tous'
+        assert messages[0][3] == ['NOTE: hello']
 
     def test_utf8_bom_with_latin_magic_comment_fails(self):
-        buf = BytesIO(codecs.BOM_UTF8 + u"""# -*- coding: latin-1 -*-
+        buf = BytesIO(codecs.BOM_UTF8 + """# -*- coding: latin-1 -*-
 # NOTE: hello
 msg = _('Bonjour à tous')
 """.encode('utf-8'))
-        self.assertRaises(SyntaxError, list,
-                          extract.extract_python(buf, ('_',), ['NOTE:'], {}))
+        with pytest.raises(SyntaxError):
+            list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
 
     def test_utf8_raw_strings_match_unicode_strings(self):
-        buf = BytesIO(codecs.BOM_UTF8 + u"""
+        buf = BytesIO(codecs.BOM_UTF8 + """
 msg = _('Bonjour à tous')
 msgu = _(u'Bonjour à tous')
 """.encode('utf-8'))
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual(u'Bonjour à tous', messages[0][2])
-        self.assertEqual(messages[0][2], messages[1][2])
+        assert messages[0][2] == 'Bonjour à tous'
+        assert messages[0][2] == messages[1][2]
 
     def test_extract_strip_comment_tags(self):
         buf = BytesIO(b"""\
@@ -417,12 +405,10 @@ _('Servus')
 _('Babatschi')""")
         messages = list(extract.extract('python', buf, comment_tags=['NOTE:', ':'],
                                         strip_comment_tags=True))
-        self.assertEqual(u'Servus', messages[0][1])
-        self.assertEqual([u'This is a comment with a very simple',
-                          u'prefix specified'], messages[0][2])
-        self.assertEqual(u'Babatschi', messages[1][1])
-        self.assertEqual([u'This is a multiline comment with',
-                          u'a prefix too'], messages[1][2])
+        assert messages[0][1] == 'Servus'
+        assert messages[0][2] == ['This is a comment with a very simple', 'prefix specified']
+        assert messages[1][1] == 'Babatschi'
+        assert messages[1][2] == ['This is a multiline comment with', 'a prefix too']
 
     def test_nested_messages(self):
         buf = BytesIO(b"""
@@ -438,22 +424,22 @@ _(u'Hello, {0} and {1}!', _(u'Heungsub'),
   _(u'Armin'))
 """)
         messages = list(extract.extract_python(buf, ('_',), ['NOTE:'], {}))
-        self.assertEqual((u'Hello, {name}!', None), messages[0][2])
-        self.assertEqual([u'NOTE: First'], messages[0][3])
-        self.assertEqual(u'Foo Bar', messages[1][2])
-        self.assertEqual([], messages[1][3])
-        self.assertEqual((u'Hello, {name1} and {name2}!', None), messages[2][2])
-        self.assertEqual([u'NOTE: Second'], messages[2][3])
-        self.assertEqual(u'Heungsub', messages[3][2])
-        self.assertEqual([], messages[3][3])
-        self.assertEqual(u'Armin', messages[4][2])
-        self.assertEqual([], messages[4][3])
-        self.assertEqual((u'Hello, {0} and {1}!', None), messages[5][2])
-        self.assertEqual([u'NOTE: Third'], messages[5][3])
-        self.assertEqual(u'Heungsub', messages[6][2])
-        self.assertEqual([], messages[6][3])
-        self.assertEqual(u'Armin', messages[7][2])
-        self.assertEqual([], messages[7][3])
+        assert messages[0][2] == ('Hello, {name}!', None)
+        assert messages[0][3] == ['NOTE: First']
+        assert messages[1][2] == 'Foo Bar'
+        assert messages[1][3] == []
+        assert messages[2][2] == ('Hello, {name1} and {name2}!', None)
+        assert messages[2][3] == ['NOTE: Second']
+        assert messages[3][2] == 'Heungsub'
+        assert messages[3][3] == []
+        assert messages[4][2] == 'Armin'
+        assert messages[4][3] == []
+        assert messages[5][2] == ('Hello, {0} and {1}!', None)
+        assert messages[5][3] == ['NOTE: Third']
+        assert messages[6][2] == 'Heungsub'
+        assert messages[6][3] == []
+        assert messages[7][2] == 'Armin'
+        assert messages[7][3] == []
 
 
 class ExtractTestCase(unittest.TestCase):
@@ -474,13 +460,16 @@ msg10 = dngettext(domain, 'Page', 'Pages', 3)
         messages = \
             list(extract.extract('python', buf, extract.DEFAULT_KEYWORDS, [],
                                  {}))
-        self.assertEqual([(5, (u'bunny', u'bunnies'), [], None),
-                          (8, u'Rabbit', [], None),
-                          (10, (u'Page', u'Pages'), [], None)], messages)
+        assert messages == [
+            (5, ('bunny', 'bunnies'), [], None),
+            (8, 'Rabbit', [], None),
+            (10, ('Page', 'Pages'), [], None),
+        ]
 
     def test_invalid_extract_method(self):
         buf = BytesIO(b'')
-        self.assertRaises(ValueError, list, extract.extract('spam', buf))
+        with pytest.raises(ValueError):
+            list(extract.extract('spam', buf))
 
     def test_different_signatures(self):
         buf = BytesIO(b"""
@@ -494,9 +483,9 @@ n = ngettext('foo')
         messages = \
             list(extract.extract('python', buf, extract.DEFAULT_KEYWORDS, [],
                                  {}))
-        self.assertEqual(len(messages), 2)
-        self.assertEqual(u'foo', messages[0][1])
-        self.assertEqual((u'hello', u'there'), messages[1][1])
+        assert len(messages) == 2
+        assert messages[0][1] == 'foo'
+        assert messages[1][1] == ('hello', 'there')
 
     def test_empty_string_msgid(self):
         buf = BytesIO(b"""\
@@ -508,7 +497,7 @@ msg = _('')
             messages = \
                 list(extract.extract('python', buf, extract.DEFAULT_KEYWORDS,
                                      [], {}))
-            self.assertEqual([], messages)
+            assert messages == []
             assert 'warning: Empty msgid.' in sys.stderr.getvalue()
         finally:
             sys.stderr = stderr
@@ -519,7 +508,7 @@ msg = _('')
         sys.stderr = StringIO()
         try:
             messages = extract.extract('python', buf)
-            self.assertEqual([], list(messages))
+            assert list(messages) == []
             assert 'warning: Empty msgid.' in sys.stderr.getvalue()
         finally:
             sys.stderr = stderr
@@ -538,4 +527,31 @@ nbsp = _('\xa0')
 """)
         messages = list(extract.extract('python', buf,
                                         extract.DEFAULT_KEYWORDS, [], {}))
-        assert messages[0][1] == u'\xa0'
+        assert messages[0][1] == '\xa0'
+
+    def test_f_strings(self):
+        buf = BytesIO(br"""
+t1 = _('foobar')
+t2 = _(f'spameggs' f'feast')  # should be extracted; constant parts only
+t2 = _(f'spameggs' 'kerroshampurilainen')  # should be extracted (mixing f with no f)
+t3 = _(f'''whoa! a '''  # should be extracted (continues on following lines)
+f'flying shark'
+    '... hello'
+)
+t4 = _(f'spameggs {t1}')  # should not be extracted
+""")
+        messages = list(extract.extract('python', buf, extract.DEFAULT_KEYWORDS, [], {}))
+        assert len(messages) == 4
+        assert messages[0][1] == 'foobar'
+        assert messages[1][1] == 'spameggsfeast'
+        assert messages[2][1] == 'spameggskerroshampurilainen'
+        assert messages[3][1] == 'whoa! a flying shark... hello'
+
+    def test_f_strings_non_utf8(self):
+        buf = BytesIO(b"""
+# -- coding: latin-1 --
+t2 = _(f'\xe5\xe4\xf6' f'\xc5\xc4\xd6')
+""")
+        messages = list(extract.extract('python', buf, extract.DEFAULT_KEYWORDS, [], {}))
+        assert len(messages) == 1
+        assert messages[0][1] == 'åäöÅÄÖ'
