@@ -1,6 +1,5 @@
-# -*- coding: utf-8 -*-
 #
-# Copyright (C) 2007-2011 Edgewall Software, 2013-2022 the Babel team
+# Copyright (C) 2007-2011 Edgewall Software, 2013-2023 the Babel team
 # All rights reserved.
 #
 # This software is licensed as described in the file LICENSE, which
@@ -11,14 +10,17 @@
 # individuals. For the exact contribution history, see the revision
 # history and logs, available at http://babel.edgewall.org/log/.
 
-from datetime import datetime
 import unittest
+from datetime import datetime
 from io import BytesIO, StringIO
 
+import pytest
+
 from babel.core import Locale
-from babel.messages.catalog import Catalog, Message
 from babel.messages import pofile
+from babel.messages.catalog import Catalog, Message
 from babel.util import FixedOffsetTimezone
+
 
 class ReadPoTestCase(unittest.TestCase):
 
@@ -26,7 +28,7 @@ class ReadPoTestCase(unittest.TestCase):
         buf = StringIO(r'''msgid "foo"
 msgstr "Voh"''')
         catalog = pofile.read_po(buf, locale='en_US')
-        self.assertEqual(Locale('en', 'US'), catalog.locale)
+        assert Locale('en', 'US') == catalog.locale
 
     def test_locale_gets_overridden_by_file(self):
         buf = StringIO(r'''
@@ -34,22 +36,22 @@ msgid ""
 msgstr ""
 "Language: en_US\n"''')
         catalog = pofile.read_po(buf, locale='de')
-        self.assertEqual(Locale('en', 'US'), catalog.locale)
+        assert Locale('en', 'US') == catalog.locale
         buf = StringIO(r'''
 msgid ""
 msgstr ""
 "Language: ko-KR\n"''')
         catalog = pofile.read_po(buf, locale='de')
-        self.assertEqual(Locale('ko', 'KR'), catalog.locale)
+        assert Locale('ko', 'KR') == catalog.locale
 
     def test_preserve_domain(self):
         buf = StringIO(r'''msgid "foo"
 msgstr "Voh"''')
         catalog = pofile.read_po(buf, domain='mydomain')
-        self.assertEqual('mydomain', catalog.domain)
+        assert catalog.domain == 'mydomain'
 
     def test_applies_specified_encoding_during_read(self):
-        buf = BytesIO(u'''
+        buf = BytesIO('''
 msgid ""
 msgstr ""
 "Project-Id-Version:  3.15\\n"
@@ -67,7 +69,18 @@ msgstr ""
 msgid "foo"
 msgstr "bär"'''.encode('iso-8859-1'))
         catalog = pofile.read_po(buf, locale='de_DE')
-        self.assertEqual(u'bär', catalog.get('foo').string)
+        assert catalog.get('foo').string == 'bär'
+
+    def test_encoding_header_read(self):
+        buf = BytesIO(b'msgid ""\nmsgstr ""\n"Content-Type: text/plain; charset=mac_roman\\n"\n')
+        catalog = pofile.read_po(buf, locale='xx_XX')
+        assert catalog.charset == 'mac_roman'
+
+    def test_plural_forms_header_parsed(self):
+        buf = BytesIO(b'msgid ""\nmsgstr ""\n"Plural-Forms: nplurals=42; plural=(n % 11);\\n"\n')
+        catalog = pofile.read_po(buf, locale='xx_XX')
+        assert catalog.plural_expr == '(n % 11)'
+        assert catalog.num_plurals == 42
 
     def test_read_multiline(self):
         buf = StringIO(r'''msgid ""
@@ -77,11 +90,12 @@ msgstr "bär"'''.encode('iso-8859-1'))
 "loop\n"
 msgstr ""''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog))
+        assert len(catalog) == 1
         message = list(catalog)[1]
-        self.assertEqual("Here's some text that\nincludesareallylongwordthat"
-                         "mightbutshouldnt throw us into an infinite loop\n",
-                         message.id)
+        assert message.id == (
+            "Here's some text that\nincludesareallylongwordthat"
+            "mightbutshouldnt throw us into an infinite loop\n"
+        )
 
     def test_fuzzy_header(self):
         buf = StringIO(r'''
@@ -94,8 +108,8 @@ msgstr ""''')
 #, fuzzy
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(list(catalog)))
-        self.assertEqual(True, list(catalog)[0].fuzzy)
+        assert len(list(catalog)) == 1
+        assert list(catalog)[0].fuzzy
 
     def test_not_fuzzy_header(self):
         buf = StringIO(r'''
@@ -107,8 +121,8 @@ msgstr ""''')
 #
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(list(catalog)))
-        self.assertEqual(False, list(catalog)[0].fuzzy)
+        assert len(list(catalog)) == 1
+        assert not list(catalog)[0].fuzzy
 
     def test_header_entry(self):
         buf = StringIO(r'''
@@ -134,18 +148,15 @@ msgstr ""
 "Generated-By: Babel 1.0dev-r313\n"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(list(catalog)))
-        self.assertEqual(u'3.15', catalog.version)
-        self.assertEqual(u'Fliegender Zirkus <fliegender@zirkus.de>',
-                         catalog.msgid_bugs_address)
-        self.assertEqual(datetime(2007, 9, 27, 11, 19,
-                                  tzinfo=FixedOffsetTimezone(7 * 60)),
-                         catalog.creation_date)
-        self.assertEqual(u'John <cleese@bavaria.de>', catalog.last_translator)
-        self.assertEqual(Locale('de'), catalog.locale)
-        self.assertEqual(u'German Lang <de@babel.org>', catalog.language_team)
-        self.assertEqual(u'iso-8859-2', catalog.charset)
-        self.assertEqual(True, list(catalog)[0].fuzzy)
+        assert len(list(catalog)) == 1
+        assert catalog.version == '3.15'
+        assert catalog.msgid_bugs_address == 'Fliegender Zirkus <fliegender@zirkus.de>'
+        assert datetime(2007, 9, 27, 11, 19, tzinfo=FixedOffsetTimezone(7 * 60)) == catalog.creation_date
+        assert catalog.last_translator == 'John <cleese@bavaria.de>'
+        assert Locale('de') == catalog.locale
+        assert catalog.language_team == 'German Lang <de@babel.org>'
+        assert catalog.charset == 'iso-8859-2'
+        assert list(catalog)[0].fuzzy
 
     def test_obsolete_message(self):
         buf = StringIO(r'''# This is an obsolete message
@@ -158,12 +169,12 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(1, len(catalog.obsolete))
-        message = catalog.obsolete[u'foo']
-        self.assertEqual(u'foo', message.id)
-        self.assertEqual(u'Voh', message.string)
-        self.assertEqual(['This is an obsolete message'], message.user_comments)
+        assert len(catalog) == 1
+        assert len(catalog.obsolete) == 1
+        message = catalog.obsolete['foo']
+        assert message.id == 'foo'
+        assert message.string == 'Voh'
+        assert message.user_comments == ['This is an obsolete message']
 
     def test_obsolete_message_ignored(self):
         buf = StringIO(r'''# This is an obsolete message
@@ -176,8 +187,8 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf, ignore_obsolete=True)
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(0, len(catalog.obsolete))
+        assert len(catalog) == 1
+        assert len(catalog.obsolete) == 0
 
     def test_multi_line_obsolete_message(self):
         buf = StringIO(r'''# This is an obsolete message
@@ -194,11 +205,11 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog.obsolete))
-        message = catalog.obsolete[u'foofoo']
-        self.assertEqual(u'foofoo', message.id)
-        self.assertEqual(u'VohVooooh', message.string)
-        self.assertEqual(['This is an obsolete message'], message.user_comments)
+        assert len(catalog.obsolete) == 1
+        message = catalog.obsolete['foofoo']
+        assert message.id == 'foofoo'
+        assert message.string == 'VohVooooh'
+        assert message.user_comments == ['This is an obsolete message']
 
     def test_unit_following_multi_line_obsolete_message(self):
         buf = StringIO(r'''# This is an obsolete message
@@ -215,11 +226,11 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog))
-        message = catalog[u'bar']
-        self.assertEqual(u'bar', message.id)
-        self.assertEqual(u'Bahr', message.string)
-        self.assertEqual(['This message is not obsolete'], message.user_comments)
+        assert len(catalog) == 1
+        message = catalog['bar']
+        assert message.id == 'bar'
+        assert message.string == 'Bahr'
+        assert message.user_comments == ['This message is not obsolete']
 
     def test_unit_before_obsolete_is_not_obsoleted(self):
         buf = StringIO(r'''
@@ -237,11 +248,11 @@ msgstr "Bahr"
 #~ "Vooooh"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog))
-        message = catalog[u'bar']
-        self.assertEqual(u'bar', message.id)
-        self.assertEqual(u'Bahr', message.string)
-        self.assertEqual(['This message is not obsolete'], message.user_comments)
+        assert len(catalog) == 1
+        message = catalog['bar']
+        assert message.id == 'bar'
+        assert message.string == 'Bahr'
+        assert message.user_comments == ['This message is not obsolete']
 
     def test_with_context(self):
         buf = BytesIO(b'''# Some string in the menu
@@ -257,17 +268,16 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf, ignore_obsolete=True)
-        self.assertEqual(2, len(catalog))
+        assert len(catalog) == 2
         message = catalog.get('foo', context='Menu')
-        self.assertEqual('Menu', message.context)
+        assert message.context == 'Menu'
         message = catalog.get('bar', context='Menu')
-        self.assertEqual('Menu', message.context)
+        assert message.context == 'Menu'
 
         # And verify it pass through write_po
         out_buf = BytesIO()
         pofile.write_po(out_buf, catalog, omit_header=True)
-        assert out_buf.getvalue().strip() == buf.getvalue().strip(), \
-            out_buf.getvalue()
+        assert out_buf.getvalue().strip() == buf.getvalue().strip()
 
     def test_obsolete_message_with_context(self):
         buf = StringIO('''
@@ -286,11 +296,11 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(2, len(catalog))
-        self.assertEqual(1, len(catalog.obsolete))
-        message = catalog.obsolete[u"foo"]
-        self.assertEqual(message.context, "other")
-        self.assertEqual(message.string, "Voh")
+        assert len(catalog) == 2
+        assert len(catalog.obsolete) == 1
+        message = catalog.obsolete["foo"]
+        assert message.context == 'other'
+        assert message.string == 'Voh'
 
     def test_multiline_context(self):
         buf = StringIO('''
@@ -301,10 +311,10 @@ msgid "mid"
 msgstr "mst"
         ''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(1, len(catalog))
+        assert len(catalog) == 1
         message = catalog.get('mid', context="a really long message context why?")
         assert message is not None
-        self.assertEqual("a really long message context why?", message.context)
+        assert message.context == 'a really long message context why?'
 
     def test_with_context_two(self):
         buf = BytesIO(b'''msgctxt "Menu"
@@ -316,11 +326,11 @@ msgid "bar"
 msgstr "Bahr"
 ''')
         catalog = pofile.read_po(buf, ignore_obsolete=True)
-        self.assertEqual(2, len(catalog))
+        assert len(catalog) == 2
         message = catalog.get('foo', context='Menu')
-        self.assertEqual('Menu', message.context)
+        assert message.context == 'Menu'
         message = catalog.get('bar', context='Mannu')
-        self.assertEqual('Mannu', message.context)
+        assert message.context == 'Mannu'
 
         # And verify it pass through write_po
         out_buf = BytesIO()
@@ -332,10 +342,10 @@ msgstr "Bahr"
 msgid_plural "foos"
 msgstr[0] "Voh"''')
         catalog = pofile.read_po(buf, locale='ja_JP')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(1, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 1
         message = catalog['foo']
-        self.assertEqual(1, len(message.string))
+        assert len(message.string) == 1
 
     def test_singular_plural_form(self):
         buf = StringIO(r'''msgid "foo"
@@ -343,10 +353,10 @@ msgid_plural "foos"
 msgstr[0] "Voh"
 msgstr[1] "Vohs"''')
         catalog = pofile.read_po(buf, locale='nl_NL')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(2, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 2
         message = catalog['foo']
-        self.assertEqual(2, len(message.string))
+        assert len(message.string) == 2
 
     def test_more_than_two_plural_forms(self):
         buf = StringIO(r'''msgid "foo"
@@ -355,11 +365,11 @@ msgstr[0] "Voh"
 msgstr[1] "Vohs"
 msgstr[2] "Vohss"''')
         catalog = pofile.read_po(buf, locale='lv_LV')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(3, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 3
         message = catalog['foo']
-        self.assertEqual(3, len(message.string))
-        self.assertEqual(u'Vohss', message.string[2])
+        assert len(message.string) == 3
+        assert message.string[2] == 'Vohss'
 
     def test_plural_with_square_brackets(self):
         buf = StringIO(r'''msgid "foo"
@@ -367,10 +377,10 @@ msgid_plural "foos"
 msgstr[0] "Voh [text]"
 msgstr[1] "Vohs [text]"''')
         catalog = pofile.read_po(buf, locale='nb_NO')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(2, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 2
         message = catalog['foo']
-        self.assertEqual(2, len(message.string))
+        assert len(message.string) == 2
 
     def test_obsolete_plural_with_square_brackets(self):
         buf = StringIO('''\
@@ -380,13 +390,13 @@ msgstr[1] "Vohs [text]"''')
 #~ msgstr[1] "Vohs [text]"
 ''')
         catalog = pofile.read_po(buf, locale='nb_NO')
-        self.assertEqual(0, len(catalog))
-        self.assertEqual(1, len(catalog.obsolete))
-        self.assertEqual(2, catalog.num_plurals)
+        assert len(catalog) == 0
+        assert len(catalog.obsolete) == 1
+        assert catalog.num_plurals == 2
         message = catalog.obsolete[('foo', 'foos')]
-        self.assertEqual(2, len(message.string))
-        self.assertEqual("Voh [text]", message.string[0])
-        self.assertEqual("Vohs [text]", message.string[1])
+        assert len(message.string) == 2
+        assert message.string[0] == 'Voh [text]'
+        assert message.string[1] == 'Vohs [text]'
 
     def test_missing_plural(self):
         buf = StringIO('''\
@@ -400,13 +410,13 @@ msgstr[0] "Voh [text]"
 msgstr[1] "Vohs [text]"
 ''')
         catalog = pofile.read_po(buf, locale='nb_NO')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(3, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 3
         message = catalog['foo']
-        self.assertEqual(3, len(message.string))
-        self.assertEqual("Voh [text]", message.string[0])
-        self.assertEqual("Vohs [text]", message.string[1])
-        self.assertEqual("", message.string[2])
+        assert len(message.string) == 3
+        assert message.string[0] == 'Voh [text]'
+        assert message.string[1] == 'Vohs [text]'
+        assert message.string[2] == ''
 
     def test_missing_plural_in_the_middle(self):
         buf = StringIO('''\
@@ -420,13 +430,13 @@ msgstr[0] "Voh [text]"
 msgstr[2] "Vohs [text]"
 ''')
         catalog = pofile.read_po(buf, locale='nb_NO')
-        self.assertEqual(1, len(catalog))
-        self.assertEqual(3, catalog.num_plurals)
+        assert len(catalog) == 1
+        assert catalog.num_plurals == 3
         message = catalog['foo']
-        self.assertEqual(3, len(message.string))
-        self.assertEqual("Voh [text]", message.string[0])
-        self.assertEqual("", message.string[1])
-        self.assertEqual("Vohs [text]", message.string[2])
+        assert len(message.string) == 3
+        assert message.string[0] == 'Voh [text]'
+        assert message.string[1] == ''
+        assert message.string[2] == 'Vohs [text]'
 
     def test_abort_invalid_po_file(self):
         invalid_po = '''
@@ -460,10 +470,8 @@ msgstr[2] "Vohs [text]"
 
         # Catalog not created, throws PoFileError
         buf = StringIO(invalid_po_2)
-        output = None
-        with self.assertRaises(pofile.PoFileError) as e:
-            output = pofile.read_po(buf, locale='fr', abort_invalid=True)
-        assert not output
+        with pytest.raises(pofile.PoFileError):
+            pofile.read_po(buf, locale='fr', abort_invalid=True)
 
         # Catalog is created with warning, no abort
         buf = StringIO(invalid_po_2)
@@ -472,17 +480,15 @@ msgstr[2] "Vohs [text]"
 
         # Catalog not created, aborted with PoFileError
         buf = StringIO(invalid_po_2)
-        output = None
-        with self.assertRaises(pofile.PoFileError) as e:
-            output = pofile.read_po(buf, locale='fr', abort_invalid=True)
-        assert not output
+        with pytest.raises(pofile.PoFileError):
+            pofile.read_po(buf, locale='fr', abort_invalid=True)
 
     def test_invalid_pofile_with_abort_flag(self):
         parser = pofile.PoFileParser(None, abort_invalid=True)
         lineno = 10
-        line = u'Algo esta mal'
+        line = 'Algo esta mal'
         msg = 'invalid file'
-        with self.assertRaises(pofile.PoFileError) as e:
+        with pytest.raises(pofile.PoFileError):
             parser._invalid_pofile(line, lineno, msg)
 
 
@@ -490,32 +496,32 @@ class WritePoTestCase(unittest.TestCase):
 
     def test_join_locations(self):
         catalog = Catalog()
-        catalog.add(u'foo', locations=[('main.py', 1)])
-        catalog.add(u'foo', locations=[('utils.py', 3)])
+        catalog.add('foo', locations=[('main.py', 1)])
+        catalog.add('foo', locations=[('utils.py', 3)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#: main.py:1 utils.py:3
+        assert buf.getvalue().strip() == b'''#: main.py:1 utils.py:3
 msgid "foo"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_write_po_file_with_specified_charset(self):
         catalog = Catalog(charset='iso-8859-1')
-        catalog.add('foo', u'äöü', locations=[('main.py', 1)])
+        catalog.add('foo', 'äöü', locations=[('main.py', 1)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=False)
         po_file = buf.getvalue().strip()
         assert b'"Content-Type: text/plain; charset=iso-8859-1\\n"' in po_file
-        assert u'msgstr "äöü"'.encode('iso-8859-1') in po_file
+        assert 'msgstr "äöü"'.encode('iso-8859-1') in po_file
 
     def test_duplicate_comments(self):
         catalog = Catalog()
-        catalog.add(u'foo', auto_comments=['A comment'])
-        catalog.add(u'foo', auto_comments=['A comment'])
+        catalog.add('foo', auto_comments=['A comment'])
+        catalog.add('foo', auto_comments=['A comment'])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#. A comment
+        assert buf.getvalue().strip() == b'''#. A comment
 msgid "foo"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_wrap_long_lines(self):
         text = """Here's some text where
@@ -529,14 +535,14 @@ not be removed
         buf = BytesIO()
         pofile.write_po(buf, catalog, no_location=True, omit_header=True,
                         width=42)
-        self.assertEqual(b'''msgid ""
+        assert buf.getvalue().strip() == b'''msgid ""
 "Here's some text where\\n"
 "white space and line breaks matter, and"
 " should\\n"
 "\\n"
 "not be removed\\n"
 "\\n"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_wrap_long_lines_with_long_word(self):
         text = """Here's some text that
@@ -547,12 +553,12 @@ includesareallylongwordthatmightbutshouldnt throw us into an infinite loop
         buf = BytesIO()
         pofile.write_po(buf, catalog, no_location=True, omit_header=True,
                         width=32)
-        self.assertEqual(b'''msgid ""
+        assert buf.getvalue().strip() == b'''msgid ""
 "Here's some text that\\n"
 "includesareallylongwordthatmightbutshouldnt"
 " throw us into an infinite "
 "loop\\n"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_wrap_long_lines_in_header(self):
         """
@@ -562,29 +568,29 @@ msgstr ""''', buf.getvalue().strip())
                           revision_date=datetime(2007, 4, 1))
         buf = BytesIO()
         pofile.write_po(buf, catalog)
-        self.assertEqual(b'''\
+        assert b'\n'.join(buf.getvalue().splitlines()[:7]) == b'''\
 # Translations template for AReallyReallyLongNameForAProject.
 # Copyright (C) 2007 ORGANIZATION
 # This file is distributed under the same license as the
 # AReallyReallyLongNameForAProject project.
 # FIRST AUTHOR <EMAIL@ADDRESS>, 2007.
 #
-#, fuzzy''', b'\n'.join(buf.getvalue().splitlines()[:7]))
+#, fuzzy'''
 
     def test_wrap_locations_with_hyphens(self):
         catalog = Catalog()
-        catalog.add(u'foo', locations=[
+        catalog.add('foo', locations=[
             ('doupy/templates/base/navmenu.inc.html.py', 60)
         ])
-        catalog.add(u'foo', locations=[
+        catalog.add('foo', locations=[
             ('doupy/templates/job-offers/helpers.html', 22)
         ])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#: doupy/templates/base/navmenu.inc.html.py:60
+        assert buf.getvalue().strip() == b'''#: doupy/templates/base/navmenu.inc.html.py:60
 #: doupy/templates/job-offers/helpers.html:22
 msgid "foo"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_no_wrap_and_width_behaviour_on_comments(self):
         catalog = Catalog()
@@ -593,7 +599,7 @@ msgstr ""''', buf.getvalue().strip())
                     locations=[("fake.py", n) for n in range(1, 30)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, width=None, omit_header=True)
-        self.assertEqual(b"""\
+        assert buf.getvalue().lower() == b"""\
 #: fake.py:1 fake.py:2 fake.py:3 fake.py:4 fake.py:5 fake.py:6 fake.py:7
 #: fake.py:8 fake.py:9 fake.py:10 fake.py:11 fake.py:12 fake.py:13 fake.py:14
 #: fake.py:15 fake.py:16 fake.py:17 fake.py:18 fake.py:19 fake.py:20 fake.py:21
@@ -602,10 +608,10 @@ msgstr ""''', buf.getvalue().strip())
 msgid "pretty dam long message id, which must really be big to test this wrap behaviour, if not it won't work."
 msgstr ""
 
-""", buf.getvalue().lower())
+"""
         buf = BytesIO()
         pofile.write_po(buf, catalog, width=100, omit_header=True)
-        self.assertEqual(b"""\
+        assert buf.getvalue().lower() == b"""\
 #: fake.py:1 fake.py:2 fake.py:3 fake.py:4 fake.py:5 fake.py:6 fake.py:7 fake.py:8 fake.py:9 fake.py:10
 #: fake.py:11 fake.py:12 fake.py:13 fake.py:14 fake.py:15 fake.py:16 fake.py:17 fake.py:18 fake.py:19
 #: fake.py:20 fake.py:21 fake.py:22 fake.py:23 fake.py:24 fake.py:25 fake.py:26 fake.py:27 fake.py:28
@@ -615,18 +621,18 @@ msgid ""
 " work."
 msgstr ""
 
-""", buf.getvalue().lower())
+"""
 
     def test_pot_with_translator_comments(self):
         catalog = Catalog()
-        catalog.add(u'foo', locations=[('main.py', 1)],
+        catalog.add('foo', locations=[('main.py', 1)],
                     auto_comments=['Comment About `foo`'])
-        catalog.add(u'bar', locations=[('utils.py', 3)],
+        catalog.add('bar', locations=[('utils.py', 3)],
                     user_comments=['Comment About `bar` with',
                                    'multiple lines.'])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#. Comment About `foo`
+        assert buf.getvalue().strip() == b'''#. Comment About `foo`
 #: main.py:1
 msgid "foo"
 msgstr ""
@@ -635,27 +641,27 @@ msgstr ""
 # multiple lines.
 #: utils.py:3
 msgid "bar"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_po_with_obsolete_message(self):
         catalog = Catalog()
-        catalog.add(u'foo', u'Voh', locations=[('main.py', 1)])
-        catalog.obsolete['bar'] = Message(u'bar', u'Bahr',
+        catalog.add('foo', 'Voh', locations=[('main.py', 1)])
+        catalog.obsolete['bar'] = Message('bar', 'Bahr',
                                           locations=[('utils.py', 3)],
                                           user_comments=['User comment'])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#: main.py:1
+        assert buf.getvalue().strip() == b'''#: main.py:1
 msgid "foo"
 msgstr "Voh"
 
 # User comment
 #~ msgid "bar"
-#~ msgstr "Bahr"''', buf.getvalue().strip())
+#~ msgstr "Bahr"'''
 
     def test_po_with_multiline_obsolete_message(self):
         catalog = Catalog()
-        catalog.add(u'foo', u'Voh', locations=[('main.py', 1)])
+        catalog.add('foo', 'Voh', locations=[('main.py', 1)])
         msgid = r"""Here's a message that covers
 multiple lines, and should still be handled
 correctly.
@@ -668,7 +674,7 @@ correctly.
                                           locations=[('utils.py', 3)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True)
-        self.assertEqual(b'''#: main.py:1
+        assert buf.getvalue().strip() == b'''#: main.py:1
 msgid "foo"
 msgstr "Voh"
 
@@ -679,51 +685,51 @@ msgstr "Voh"
 #~ msgstr ""
 #~ "Here's a message that covers\\n"
 #~ "multiple lines, and should still be handled\\n"
-#~ "correctly.\\n"''', buf.getvalue().strip())
+#~ "correctly.\\n"'''
 
     def test_po_with_obsolete_message_ignored(self):
         catalog = Catalog()
-        catalog.add(u'foo', u'Voh', locations=[('main.py', 1)])
-        catalog.obsolete['bar'] = Message(u'bar', u'Bahr',
+        catalog.add('foo', 'Voh', locations=[('main.py', 1)])
+        catalog.obsolete['bar'] = Message('bar', 'Bahr',
                                           locations=[('utils.py', 3)],
                                           user_comments=['User comment'])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True, ignore_obsolete=True)
-        self.assertEqual(b'''#: main.py:1
+        assert buf.getvalue().strip() == b'''#: main.py:1
 msgid "foo"
-msgstr "Voh"''', buf.getvalue().strip())
+msgstr "Voh"'''
 
     def test_po_with_previous_msgid(self):
         catalog = Catalog()
-        catalog.add(u'foo', u'Voh', locations=[('main.py', 1)],
-                    previous_id=u'fo')
+        catalog.add('foo', 'Voh', locations=[('main.py', 1)],
+                    previous_id='fo')
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True, include_previous=True)
-        self.assertEqual(b'''#: main.py:1
+        assert buf.getvalue().strip() == b'''#: main.py:1
 #| msgid "fo"
 msgid "foo"
-msgstr "Voh"''', buf.getvalue().strip())
+msgstr "Voh"'''
 
     def test_po_with_previous_msgid_plural(self):
         catalog = Catalog()
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'),
-                    locations=[('main.py', 1)], previous_id=(u'fo', u'fos'))
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'),
+                    locations=[('main.py', 1)], previous_id=('fo', 'fos'))
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True, include_previous=True)
-        self.assertEqual(b'''#: main.py:1
+        assert buf.getvalue().strip() == b'''#: main.py:1
 #| msgid "fo"
 #| msgid_plural "fos"
 msgid "foo"
 msgid_plural "foos"
 msgstr[0] "Voh"
-msgstr[1] "Voeh"''', buf.getvalue().strip())
+msgstr[1] "Voeh"'''
 
     def test_sorted_po(self):
         catalog = Catalog()
-        catalog.add(u'bar', locations=[('utils.py', 3)],
+        catalog.add('bar', locations=[('utils.py', 3)],
                     user_comments=['Comment About `bar` with',
                                    'multiple lines.'])
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'),
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'),
                     locations=[('main.py', 1)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, sort_output=True)
@@ -744,12 +750,12 @@ msgstr[1] "Voeh"''' in value
 
     def test_sorted_po_context(self):
         catalog = Catalog()
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'),
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'),
                     locations=[('main.py', 1)],
                     context='there')
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'),
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'),
                     locations=[('main.py', 1)])
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'),
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'),
                     locations=[('main.py', 1)],
                     context='here')
         buf = BytesIO()
@@ -779,8 +785,8 @@ msgstr[1] "Voeh"''' in value
 
     def test_file_sorted_po(self):
         catalog = Catalog()
-        catalog.add(u'bar', locations=[('utils.py', 3)])
-        catalog.add((u'foo', u'foos'), (u'Voh', u'Voeh'), locations=[('main.py', 1)])
+        catalog.add('bar', locations=[('utils.py', 3)])
+        catalog.add(('foo', 'foos'), ('Voh', 'Voeh'), locations=[('main.py', 1)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, sort_by_file=True)
         value = buf.getvalue().strip()
@@ -788,7 +794,7 @@ msgstr[1] "Voeh"''' in value
 
     def test_file_with_no_lineno(self):
         catalog = Catalog()
-        catalog.add(u'bar', locations=[('utils.py', None)],
+        catalog.add('bar', locations=[('utils.py', None)],
                     user_comments=['Comment About `bar` with',
                                    'multiple lines.'])
         buf = BytesIO()
@@ -811,52 +817,51 @@ msgstr ""
 msgid "broken line number"
 msgstr ""''')
         catalog = pofile.read_po(buf)
-        self.assertEqual(catalog['missing line number'].locations, [(u'broken_file.py', None)])
-        self.assertEqual(catalog['broken line number'].locations, [])
+        assert catalog['missing line number'].locations == [('broken_file.py', None)]
+        assert catalog['broken line number'].locations == []
 
     def test_include_lineno(self):
         catalog = Catalog()
-        catalog.add(u'foo', locations=[('main.py', 1)])
-        catalog.add(u'foo', locations=[('utils.py', 3)])
+        catalog.add('foo', locations=[('main.py', 1)])
+        catalog.add('foo', locations=[('utils.py', 3)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True, include_lineno=True)
-        self.assertEqual(b'''#: main.py:1 utils.py:3
+        assert buf.getvalue().strip() == b'''#: main.py:1 utils.py:3
 msgid "foo"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
     def test_no_include_lineno(self):
         catalog = Catalog()
-        catalog.add(u'foo', locations=[('main.py', 1)])
-        catalog.add(u'foo', locations=[('main.py', 2)])
-        catalog.add(u'foo', locations=[('utils.py', 3)])
+        catalog.add('foo', locations=[('main.py', 1)])
+        catalog.add('foo', locations=[('main.py', 2)])
+        catalog.add('foo', locations=[('utils.py', 3)])
         buf = BytesIO()
         pofile.write_po(buf, catalog, omit_header=True, include_lineno=False)
-        self.assertEqual(b'''#: main.py utils.py
+        assert buf.getvalue().strip() == b'''#: main.py utils.py
 msgid "foo"
-msgstr ""''', buf.getvalue().strip())
+msgstr ""'''
 
 
 class PofileFunctionsTestCase(unittest.TestCase):
 
     def test_unescape(self):
-        escaped = u'"Say:\\n  \\"hello, world!\\"\\n"'
-        unescaped = u'Say:\n  "hello, world!"\n'
-        self.assertNotEqual(unescaped, escaped)
-        self.assertEqual(unescaped, pofile.unescape(escaped))
+        escaped = '"Say:\\n  \\"hello, world!\\"\\n"'
+        unescaped = 'Say:\n  "hello, world!"\n'
+        assert unescaped != escaped
+        assert unescaped == pofile.unescape(escaped)
 
     def test_unescape_of_quoted_newline(self):
         # regression test for #198
-        self.assertEqual(r'\n', pofile.unescape(r'"\\n"'))
+        assert pofile.unescape(r'"\\n"') == '\\n'
 
     def test_denormalize_on_msgstr_without_empty_first_line(self):
         # handle irregular multi-line msgstr (no "" as first line)
         # gracefully (#171)
         msgstr = '"multi-line\\n"\n" translation"'
-        expected_denormalized = u'multi-line\n translation'
+        expected_denormalized = 'multi-line\n translation'
 
-        self.assertEqual(expected_denormalized, pofile.denormalize(msgstr))
-        self.assertEqual(expected_denormalized,
-                         pofile.denormalize('""\n' + msgstr))
+        assert expected_denormalized == pofile.denormalize(msgstr)
+        assert expected_denormalized == pofile.denormalize(f'""\n{msgstr}')
 
 
 def test_unknown_language_roundtrip():
