@@ -1,8 +1,8 @@
-# -- encoding: UTF-8 --
 from datetime import time
 
-import babel.dates as dates
 import pytest
+
+import babel.dates as dates
 
 
 @pytest.mark.parametrize("locale, time, expected_period_id", [
@@ -10,6 +10,9 @@ import pytest
     ("de", time(3, 11), "night1"),  # (after, before)
     ("fi", time(0), "midnight"),  # (at)
     ("en_US", time(12), "noon"),  # (at)
+    ("en_US", time(21), "night1"),  # (from, before) across 0:00
+    ("en_US", time(5), "night1"),  # (from, before) across 0:00
+    ("en_US", time(6), "morning1"),  # (from, before)
     ("agq", time(10), "am"),  # no periods defined
     ("agq", time(22), "pm"),  # no periods defined
     ("am", time(14), "afternoon1"),  # (before, after)
