@@ -1,4 +1,3 @@
-# coding=utf-8
 import pytest
 
 from babel import lists
@@ -7,11 +6,11 @@ from babel import lists
 def test_format_list():
     for list, locale, expected in [
         ([], 'en', ''),
-        ([u'string'], 'en', u'string'),
-        (['string1', 'string2'], 'en', u'string1 and string2'),
-        (['string1', 'string2', 'string3'], 'en', u'string1, string2, and string3'),
-        (['string1', 'string2', 'string3'], 'zh', u'string1、string2和string3'),
-        (['string1', 'string2', 'string3', 'string4'], 'ne', u'string1,string2, string3 र string4'),
+        (['string'], 'en', 'string'),
+        (['string1', 'string2'], 'en', 'string1 and string2'),
+        (['string1', 'string2', 'string3'], 'en', 'string1, string2, and string3'),
+        (['string1', 'string2', 'string3'], 'zh', 'string1、string2和string3'),
+        (['string1', 'string2', 'string3', 'string4'], 'ne', 'string1,string2, string3 र string4'),
     ]:
         assert lists.format_list(list, locale=locale) == expected
 
